pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "threads-cli"
version = "0.1.0"
description = "CLI for the official Meta Threads API (graph.threads.net)"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Fioenix" }]
keywords = ["threads", "meta", "api", "cli"]
dependencies = [
  "requests>=2.31",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "ruff>=0.6",
  "build>=1.2",
]

[project.scripts]
threads = "threads_cli.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
addopts = "-q"
